route

+++ backend/usecase/delete_shelter.go
@@ -10,7 +10,7 @@ import (
 )
 
 // DeleteShelter deletes a shelter by ID
-func DeleteShelter(ctx context.Context, input input.ShelterDelete) (int, error) {
+func DeleteShelter(ctx context.Context, input input.ShelterDelete, userID string, userRole string) (int, error) {
 	id, err := strconv.Atoi(input.ID)
 	if err != nil {
 		return http.StatusBadRequest, err
@@ -19,6 +19,15 @@ func DeleteShelter(ctx context.Context, input input.ShelterDelete) (int, error) 
 	shelterRepo, err := db.NewShelterRepo(ctx)
 	if err != nil {
 		return http.StatusInternalServerError, err
+	}
+
+	existingShelter, _, err := shelterRepo.Get(ctx, id)
+	if existingShelter == nil {
+		return http.StatusNotFound, nil
+	}
+	// 認可判定
+	if !(userRole == "admin" || (userRole == "editor" && existingShelter.UserID == userID)) {
+		return http.StatusForbidden, nil
 	}
 
 	statusCode, err := shelterRepo.Delete(ctx, id)
